% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavplot.R
\name{draw.arrow}
\alias{draw.arrow}
\title{draw.arrow}
\usage{
draw.arrow(x0, y0, x1, y1, cex = 0.2)
}
\arguments{
\item{x0}{Starting x coord}

\item{y0}{Starting y coord}

\item{x1}{Ending x coord}

\item{y1}{Ending x coord}

\item{cex}{Size of node (for spacing purposes).}
}
\value{
Adds edge with closed arrow to plot.
}
\description{
draw.arrow
}

