\name{lavplot-package}
\alias{lavplot-package}
\docType{package}
\title{
\packageTitle{lavplot}
}
\description{
\if{html}{\figure{logo.svg}{options: width=60 alt="lavplot logo" style="float:left;"}}
\if{latex}{\figure{logo.pdf}{options: width=0.5in}}
\packageDescription{lavplot}
}
\author{
\packageAuthor{lavplot}

Maintainer: \packageMaintainer{lavplot}
}
\keyword{package}
