% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lavplot.R
\name{plot_nodes}
\alias{plot_nodes}
\title{plot_nodes}
\usage{
plot_nodes(x, y, cex = 0.2, a = 13, h = 1, w = 1)
}
\arguments{
\item{x}{Vector of x coords with recommended min/max of
-.75 to .75.}

\item{y}{Vector of y coords with recommended min/max of
-.75 to .75.}

\item{cex}{Size of nodes defaults to .2}

\item{a}{Something to do with color manipulation.}

\item{h}{height of plot defaults to 1.}

\item{w}{width of plot defaults to 1.}
}
\value{
Node diagram
}
\description{
plot_nodes
}
\examples{
cex <- .25
plot_nodes(c(-.75, -.75, -.75, 0,  .75, .75),
           c(-.75,    0,  .75, 0, -.5,  .5),
           cex = cex)
draw.arrow(-.75, -.75, 0, 0, cex)
draw.arrow(-.75, 0, 0, 0, cex)
draw.arrow(-.75, .75, 0, 0, cex)
draw.arrow(   0, 0, .75, .5, cex)
draw.arrow(   0, 0, .75, -.5, cex)
text(-.75, .75, "Education")
text(-.75, 0, "Age")
text(-.75, -.75, "Race")
text(0, 0, "Political\\nEngagement")
text(.75, .5, "Political\\nParticipation")
text(.75, -.5, "Civic\\nEngagement")

## mediation model
cex <- .35
plot_nodes(c(-1.5, 0, 1.5), c(-.5, .5, -.5),
           w = 1.5, h = 1, cex)
draw.arrow(-1.5, -.5, 1.5, -.5, cex)
draw.arrow(-1.5, -.5,  .0 ,  .5, cex)
draw.arrow( 0, .5,   1.5,  -.5, cex)
text(-1.5, -.5, "x")
text(0, .5, "w")
text(1.5, -.5, "y")
}

